Simplex descent for quadratics.
Write the problem in barycentric coordinates in closed form if the objective function is quadratic. Need to verify that the code works well for an active set with dense atoms (right now all my testing is with MaybeHotVectors), and perform more tests with the code.

Next step is to implement Nesterov's AGD.